Spaceship moves!

+++ rust-backend/src/superstellar/constants.rs
@@ -7,7 +7,7 @@ pub const PHYSICS_FRAME_DURATION: Duration = Duration::from_millis(20);
 pub const UPDATE_SEND_INTERVAL: Duration = Duration::from_millis(10);
 
 // Acceleration is spaceship's linear acceleration on thruster.
-pub const SPACESHIP_ACCELERATION: f64 = 50.0;
+pub const SPACESHIP_ACCELERATION: f32 = 50.0;
 
 // maximum speed of each Spaceship
 pub const SPACESHIP_MAX_SPEED: f64 = 600.0;
